// src/components/organism/panel/components/info/form-update-info-type.ts
export interface InitialType {
  name: string;
  last_name: string;
  gender: string;
  career: string;
  birthday: string | number;
  day_of_birth: string | number;
  month_of_birth: string | number;
  year_of_birth: string | number;
  national_code: string;
  email: string;
  postal_code: string;
  address: string;
}